# Improve performance for row and array blocks without nulls