--- a/presto-common/src/main/java/com/facebook/presto/common/block/ArrayBlock.java
+++ b/presto-common/src/main/java/com/facebook/presto/common/block/ArrayBlock.java
@@ -43,9 +43,10 @@ public class ArrayBlock
      * Create an array block directly from columnar nulls, values, and offsets into the values.
      * A null array must have no entries.
      */
-    public static Block fromElementBlock(int positionCount, Optional<boolean[]> valueIsNull, int[] arrayOffset, Block values)
+    public static Block fromElementBlock(int positionCount, Optional<boolean[]> valueIsNullOptional, int[] arrayOffset, Block values)
     {
-        validateConstructorArguments(0, positionCount, valueIsNull.orElse(null), arrayOffset, values);
+        boolean[] valueIsNull = valueIsNullOptional.orElse(null);
+        validateConstructorArguments(0, positionCount, valueIsNull, arrayOffset, values);
         // for performance reasons per element checks are only performed on the public construction
         for (int i = 0; i < positionCount; i++) {
             int offset = arrayOffset[i];
@@ -53,11 +54,11 @@ public class ArrayBlock
             if (length < 0) {
                 throw new IllegalArgumentException(format("Offset is not monotonically ascending. offsets[%s]=%s, offsets[%s]=%s", i, arrayOffset[i], i + 1, arrayOffset[i + 1]));
             }
-            if (valueIsNull.isPresent() && valueIsNull.get()[i] && length != 0) {
+            if (valueIsNull != null && valueIsNull[i] && length != 0) {
                 throw new IllegalArgumentException("A null array must have zero entries");
             }
         }
-        return new ArrayBlock(0, positionCount, valueIsNull.orElse(null), arrayOffset, values);
+        return new ArrayBlock(0, positionCount, valueIsNull, arrayOffset, values);
     }
 
     /**
@@ -185,6 +186,19 @@ public class ArrayBlock
     protected boolean[] getValueIsNull()
     {
         return valueIsNull;
+    }
+
+    @Override
+    public boolean mayHaveNull()
+    {
+        return valueIsNull != null;
+    }
+
+    @Override
+    public boolean isNull(int position)
+    {
+        checkReadablePosition(position);
+        return valueIsNull != null && valueIsNull[position + arrayOffset];
     }
 
     @Override