BenchmarkBlocks.benchmarkCopyPositions result summary for `ARRAY(ARRAY(BIGINT))` (full results of the latest iteration available at [jmh.morethan.io](https://jmh.morethan.io/?sources=https://gist.githubusercontent.com/pettyjamesm/816ad333656bab1cce4d9cb572f1b467/raw/1facbc25334162782bf7f013d2a042bee059486c/baseline_blocks.json,https://gist.githubusercontent.com/pettyjamesm/816ad333656bab1cce4d9cb572f1b467/raw/077dd152d3daebe49e62cf583e7c542b14535df6/improved_blocks_v2.json))

BEFORE
```
Benchmark                               (encodings)  (nestedNullRate)  (positionCount)  (primitiveNullRate)       (typeSignature)  (useBlockView)  Mode  Cnt      Score     Error  Units
BenchmarkBlocks.benchmarkCopyPositions         NONE              0.0f              256                 0.0f  ARRAY(ARRAY(BIGINT))            true  avgt    8  21687.932 ± 330.720  ns/op
BenchmarkBlocks.benchmarkCopyPositions         NONE              0.0f              256                 0.5f  ARRAY(ARRAY(BIGINT))            true  avgt    8  22573.464 ± 137.285  ns/op
BenchmarkBlocks.benchmarkCopyPositions         NONE             0.25f              256                 0.0f  ARRAY(ARRAY(BIGINT))            true  avgt    8  13351.485 ±  59.630  ns/op
BenchmarkBlocks.benchmarkCopyPositions         NONE             0.25f              256                 0.5f  ARRAY(ARRAY(BIGINT))            true  avgt    8  14558.172 ± 409.044  ns/op
BenchmarkBlocks.benchmarkCopyPositions         NONE              0.5f              256                 0.0f  ARRAY(ARRAY(BIGINT))            true  avgt    8   9559.983 ± 239.692  ns/op
BenchmarkBlocks.benchmarkCopyPositions         NONE              0.5f              256                 0.5f  ARRAY(ARRAY(BIGINT))            true  avgt    8   8485.457 ±  40.062  ns/op
BenchmarkBlocks.benchmarkCopyPositions         NONE             0.75f              256                 0.0f  ARRAY(ARRAY(BIGINT))            true  avgt    8   4256.374 ± 137.984  ns/op
BenchmarkBlocks.benchmarkCopyPositions         NONE             0.75f              256                 0.5f  ARRAY(ARRAY(BIGINT))            true  avgt    8   4110.957 ± 110.467  ns/op
```

AFTER
```
Benchmark                               (encodings)  (nestedNullRate)  (positionCount)  (primitiveNullRate)       (typeSignature)  (useBlockView)  Mode  Cnt      Score     Error  Units
BenchmarkBlocks.benchmarkCopyPositions         NONE              0.0f              256                 0.0f  ARRAY(ARRAY(BIGINT))            true  avgt    8  11823.971 ±  47.903  ns/op
BenchmarkBlocks.benchmarkCopyPositions         NONE              0.0f              256                 0.5f  ARRAY(ARRAY(BIGINT))            true  avgt    8  13759.607 ±  93.183  ns/op
BenchmarkBlocks.benchmarkCopyPositions         NONE             0.25f              256                 0.0f  ARRAY(ARRAY(BIGINT))            true  avgt    8   8674.597 ±  51.881  ns/op
BenchmarkBlocks.benchmarkCopyPositions         NONE             0.25f              256                 0.5f  ARRAY(ARRAY(BIGINT))            true  avgt    8   9391.782 ±  38.823  ns/op
BenchmarkBlocks.benchmarkCopyPositions         NONE              0.5f              256                 0.0f  ARRAY(ARRAY(BIGINT))            true  avgt    8   5687.830 ±  42.380  ns/op
BenchmarkBlocks.benchmarkCopyPositions         NONE              0.5f              256                 0.5f  ARRAY(ARRAY(BIGINT))            true  avgt    8   6138.722 ± 163.047  ns/op
BenchmarkBlocks.benchmarkCopyPositions         NONE             0.75f              256                 0.0f  ARRAY(ARRAY(BIGINT))            true  avgt    8   2438.037 ±  68.450  ns/op
BenchmarkBlocks.benchmarkCopyPositions         NONE             0.75f              256                 0.5f  ARRAY(ARRAY(BIGINT))            true  avgt    8   2568.161 ±  74.092  ns/op
```

```
== NO RELEASE NOTE ==
```