@pettyjamesm 2021-11-02
> would you include the before and after numbers for BenchmarkBlocks in the commit messages and PR message

@yingsu00 sure, I've updated the PR and commit messages to include a single benchmark run result with the latest full results linked in the PR description.