@mbasmanova 2021-11-22
@pettyjamesm @yingsu00 We are seeing a regression that might have been caused by this change. We do not have a repro that we can share publicly yet, but would appreciate if you could take a look at the code to see where the bug might be.

```
Caused by: java.lang.IndexOutOfBoundsException: Invalid position 255 and length 0 in block with 198 positions
	at com.facebook.presto.common.block.BlockUtil.checkValidRegion(BlockUtil.java:61)
	at com.facebook.presto.common.block.VariableWidthBlock.copyRegion(VariableWidthBlock.java:208)
	at com.facebook.presto.common.block.AbstractArrayBlock.copyPositions(AbstractArrayBlock.java:78)
	at com.facebook.presto.common.Page.copyPositions(Page.java:373)
	at com.facebook.presto.operator.exchange.PartitioningExchanger.accept(PartitioningExchanger.java:81)
	at com.facebook.presto.operator.exchange.LocalExchangeSink.addPage(LocalExchangeSink.java:74)
	at com.facebook.presto.operator.exchange.LocalExchangeSinkOperator.addInput(LocalExchangeSinkOperator.java:152)
	at com.facebook.presto.operator.Driver.processInternal(Driver.java:434)
	at com.facebook.presto.operator.Driver.lambda$processFor$9(Driver.java:307)
	at com.facebook.presto.operator.Driver.tryWithLock(Driver.java:728)
	at com.facebook.presto.operator.Driver.processFor(Driver.java:300)
	at com.facebook.presto.execution.SqlTaskExecution$DriverSplitRunner.processFor(SqlTaskExecution.java:1077)
	at com.facebook.presto.execution.executor.PrioritizedSplitRunner.process(PrioritizedSplitRunner.java:162)
	at com.facebook.presto.execution.executor.TaskExecutor$TaskRunner.run(TaskExecutor.java:599)
```